Remove MatchStr
