+++ crates/cgp-field/src/traits/mod.rs
@@ -10,7 +10,6 @@ mod map_field;
 mod map_fields;
 mod map_type;
 mod map_type_ref;
-mod match_str;
 mod nat;
 mod partial_data;
 mod static_format;
@@ -32,7 +31,6 @@ pub use map_field::*;
 pub use map_fields::*;
 pub use map_type::*;
 pub use map_type_ref::*;
-pub use match_str::*;
 pub use nat::*;
 pub use partial_data::*;
 pub use static_format::*;
