+++ crates/cgp-tests/src/tests/mod.rs
@@ -10,6 +10,5 @@ pub mod has_field;
 pub mod has_fields;
 pub mod monad;
 pub mod preset;
-pub mod static_string;
 pub mod symbol;
 pub mod use_delegate;
